+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - refactor: move the common code for manipulating the signature of the wrapped functions in `WithStore` and `Autorun` to a utility function
 - feat: support `with_state` to be applied to methods of classes, not just functions
 - feat: support `view` to be applied to methods of classes, not just functions, it works for `autorun` too, but only when it is being called directly like a view
+- refactor: rename `_id` field of combine reducer state to `combine_reducers_id`
 
 ## Version 0.23.0